Change 'zsh' default theme to personalised one
Since the 'danthe' personalised theme was created, it needs to be
applied.
Replace the 'robbyrussell' theme with 'danthe' to suit my needs, but
remember when pulling the repo to place the 'custom' directory in
'~/.oh-my-zsh' before replacing the '.zshrc' config file, otherwise it
will throw errors that the theme and plugin don't exist.