Use dynamic users in dotfiles paths
Dotfiles worked only for the default 'daniel' username thus throwing an
error everytime they were copied in the '~/' directory if username was
different.
If someday I decided to clone these in a VM I had to modify all Paths to
suit the Machine's username...

Replace 'daniel' username with the '$USER' global variable so that
dotfiles would function properly.

As a result, '.bashrc' and '.zshrc' will not require the user to modify
the default Paths.