Fix color display
The selected text wasn't colored or it was coloring the whole line.

As a result, call `echo` with `-e` flag to display certain sections of
text in yellow color for 'deprecated' and set the other text to normal
color.
This way coloring the whole line in yellow will be avoided.